--- a/src/Xamarin.Android.Build.Tasks/Utilities/LlvmIrGenerator/LlvmIrGenerator.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/LlvmIrGenerator/LlvmIrGenerator.cs
@@ -537,7 +537,6 @@ namespace Xamarin.Android.Tasks.LLVMIR
 			Output.Write (info.NativeTypeDesignator);
 			Output.Write ('.');
 			Output.Write (info.Name);
-
 			if (isArrayOfPointers)
 				Output.Write ('*');
 			Output.Write ("] zeroinitializer");