Reviewing `dotnet trace` output for a `dotnet new maui` app:

    dotnet trace collect --format speedscope -- .\bin\Release\dotnet\dotnet.exe build -bl --no-restore foo.csproj

21% of the time in `<GenerateJavaStubs/>` is spent in:

    249.60ms xamarin.android.build.tasks!Xamarin.Android.Tasks.LLVMIR.LlvmIrComposer.Write(value class Xamarin.Android.Tools.Andro...

Drilling down, some of the hot spots are:

    68.64ms xamarin.android.build.tasks!Xamarin.Android.Tasks.LLVMIR.LlvmIrGenerator.WriteString(class System.String,class System.St
    40.39ms xamarin.android.build.tasks!Xamarin.Android.Tasks.LLVMIR.LlvmIrGenerator.WritePointer(class Xamarin.Android.Tasks.LLVMIR

These seem to be places we create intermediate `string` or `byte[]` objects.

The types of things I solved so far:

* I audited every `$""` to see if we could call `TextWriter.Write()` or `StringBuilder.Append()` multiple times to avoid creating a string.

* Use `ArrayPool` in combination with `Encoding.GetBytes()`.

* Use `char` overloads, where we were using a 1-length `string`.

Hopefully, I've changed no behavior at all here, and we've just saved a lot of `string` and `byte[]` allocations.

After these changes go in, we can probably profile and find other areas to improve.

Testing the same `dotnet new maui` app, the overall time is now:

    190.97ms xamarin.android.build.tasks!Xamarin.Android.Tasks.LLVMIR.LlvmIrComposer.Write(value class Xamarin.Android.Tools.AndroidT

I think this will save ~60ms on incremental builds of the .NET MAUI project template. Larger projects with more Java subclasses could save even more.