@jonathanpeppers 2022-12-09
> The changes make code slightly less readable, but savings are large enough to justify it.

I like `$""` better too in most of these places...