@grendello 2022-12-09
> > The changes make code slightly less readable, but savings are large enough to justify it.
> 
> I like `$""` better too in most of these places...

The main reason I used it was to be able to see what syntax is generated by the line. I wonder, maybe a comment with target syntax above the relevant code would be a good idea?